Moved non-Lambda specific explorer nodes to shared folder
