+++ src/lambda/explorer/standaloneNodes.ts
@@ -11,9 +11,9 @@ import * as vscode from 'vscode'
 import { LambdaClient } from '../../shared/clients/lambdaClient'
 import { ext } from '../../shared/extensionGlobals'
 import { AWSTreeErrorHandlerNode } from '../../shared/treeview/awsTreeErrorHandlerNode'
+import { ErrorNode } from '../../shared/treeview/errorNode'
 import { toArrayAsync, toMap, updateInPlace } from '../../shared/utilities/collectionUtils'
 import { listLambdaFunctions } from '../utils'
-import { ErrorNode } from './errorNode'
 import { FunctionNodeBase } from './functionNode'
 import { RegionNode } from './regionNode'
