+++ src/lambda/lambdaTreeDataProvider.ts
@@ -14,6 +14,7 @@ import { defaultMetricDatum, registerCommand } from '../shared/telemetry/telemet
 import { AWSCommandTreeNode } from '../shared/treeview/awsCommandTreeNode'
 import { AWSTreeNodeBase } from '../shared/treeview/awsTreeNodeBase'
 import { RefreshableAwsTreeProvider } from '../shared/treeview/awsTreeProvider'
+import { ErrorNode } from '../shared/treeview/errorNode'
 import { intersection, toMap, updateInPlace } from '../shared/utilities/collectionUtils'
 import { ChannelLogger, localize } from '../shared/utilities/vsCodeUtils'
 import {
@@ -28,7 +29,6 @@ import { invokeLambda } from './commands/invokeLambda'
 import { showErrorDetails } from './commands/showErrorDetails'
 import { CloudFormationStackNode } from './explorer/cloudFormationNodes'
 import { DefaultRegionNode } from './explorer/defaultRegionNode'
-import { ErrorNode } from './explorer/errorNode'
 import { FunctionNodeBase } from './explorer/functionNode'
 import { RegionNode } from './explorer/regionNode'
 import { StandaloneFunctionNode } from './explorer/standaloneNodes'
