+++ src/shared/treeview/placeholderNode.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { AWSTreeNodeBase } from '../../shared/treeview/awsTreeNodeBase'
+import { AWSTreeNodeBase } from './awsTreeNodeBase'
 
 // Can be used to add a child node in an explorer when a region has no resources
 // relevant to the explorer type.
