+++ src/test/lambda/explorer/standaloneNodes.test.ts
@@ -8,7 +8,6 @@ import { Lambda } from 'aws-sdk'
 import * as os from 'os'
 import { TreeItem, Uri } from 'vscode'
 import { DefaultRegionNode } from '../../../lambda/explorer/defaultRegionNode'
-import { ErrorNode } from '../../../lambda/explorer/errorNode'
 import {
     DefaultStandaloneFunctionGroupNode,
     DefaultStandaloneFunctionNode,
@@ -21,6 +20,7 @@ import { StsClient } from '../../../shared/clients/stsClient'
 import { ext } from '../../../shared/extensionGlobals'
 import { TestLogger } from '../../../shared/loggerUtils'
 import { RegionInfo } from '../../../shared/regions/regionInfo'
+import { ErrorNode } from '../../../shared/treeview/errorNode'
 import { MockLambdaClient } from '../../shared/clients/mockClients'
 
 // TODO : Consolidate all asyncGenerator calls into a shared utility method
